chore: cargo clippy --fix + CHANGELOG addendum for deployment-feedback hardening
Two cleanup items in one commit (mechanical, no logic changes):

1. `cargo clippy --fix --workspace --all-targets --all-features` —
   38 files, +553/-538 lines. Mostly `uninlined_format_args`
   (`format!("{}", x)` → `format!("{x}")`), `collapsible_if`,
   `&Vec` → `&[]`, and `Clone` on `Copy` types. Workspace clippy
   warnings drop from 230 to 28 (the remaining 28 are structural
   — `dead_code`, derived-impl candidates — that need human
   judgement).

2. CHANGELOG addendum on the existing 0.14.0 entry covering the
   pluggable storage abstraction and the post-PR
   deployment-feedback hardening (commits c47ae38..4282ebf).
   Mirrors what's now in PR #286's body so the published
   changelog matches what reviewers see.

`cargo test --workspace` and `cargo clippy --workspace` still
green after the fixes.

+++ crates/messaging/affinidi-messaging-mediator/src/server.rs
@@ -504,7 +504,7 @@ pub async fn serve_internal(
                 .make_span_with(trace::DefaultMakeSpan::new().level(Level::INFO))
                 .on_response(trace::DefaultOnResponse::new().level(Level::INFO)),
         )
-        .layer(RequestBodyLimitLayer::new(config.limits.http_size as usize))
+        .layer(RequestBodyLimitLayer::new(config.limits.http_size))
         .layer(RateLimitLayer::new(rate_limiter))
         .layer(RequestIdLayer::new())
         .route(
